IMB-64 Remove ability to save CSV direct to DB
This could be a useful feature, but having it as an option is hard to allow/ turn on without there being a lot of risk of the process being triggered when it shouldnt be.
For the good it would do, the risk might be too great.
Rather than this I suggest we maintain our own list - adding plenty of new users as needed and making use of the cron job as a test